Add a base state from which the two different states should be subclassed.
