+++ practice.py
@@ -1,4 +1,4 @@
-import practiceGui
+import practiceGui, practiceState
 
 state = practiceState.State()
 gui = practiceGui.Gui(state)
